+++ module2/2.1/main.c
@@ -157,7 +157,7 @@ void print_address(const address *a)
 
 void print_contact(const contact *c)
 {
-    printf("Контакт #%llu:\n", c->id);
+    printf("Контакт #%lu:\n", c->id);
 
     print_names(&c->names);
     print_address(&c->address);
@@ -326,8 +326,8 @@ void menu2(phonebook *pb)
     }
 
     printf("Введите id контакта для удаления: ");
-    unsigned long long id;
-    scanf("%llu", &id);
+    uint64_t id;
+    scanf("%lu", &id);
 
     if (delete_contact_from_phonebook(pb, id) == 0)
     {
@@ -348,8 +348,8 @@ void menu3(phonebook *pb)
     }
 
     printf("Введите id контакта для редактирования: ");
-    unsigned long long id;
-    scanf("%llu", &id);
+    uint64_t id;
+    scanf("%lu", &id);
 
     printf("Форматные строки для редактирования:\n");
     printf("1.1 - фамилия, 1.2 - имя, 1.3 - отчество\n");
@@ -402,7 +402,7 @@ void menu4(phonebook *pb)
     }
     else
     {
-        for (unsigned int i = 0; i < pb->contacts_quan; i++)
+        for (uint64_t i = 0; i < pb->contacts_quan; i++)
         {
             print_contact(&(pb->contacts[i]));
         }
@@ -418,8 +418,8 @@ void menu5(phonebook *pb)
     }
 
     printf("Введите id контакта для добавления ссылки: ");
-    unsigned long long id;
-    scanf("%llu", &id);
+    uint64_t id;
+    scanf("%lu", &id);
 
     contact *c = find_by_id(pb, id);
 
@@ -456,8 +456,8 @@ void menu6(phonebook *pb)
     }
 
     printf("Введите id контакта для добавления почты: ");
-    unsigned long long id;
-    scanf("%llu", &id);
+    uint64_t id;
+    scanf("%lu", &id);
 
     contact *c = find_by_id(pb, id);
 
@@ -493,8 +493,8 @@ void menu7(phonebook *pb)
     }
 
     printf("Введите id контакта для добавления телефона: ");
-    unsigned long long id;
-    scanf("%llu", &id);
+    uint64_t id;
+    scanf("%lu", &id);
 
     contact *c = find_by_id(pb, id);
 
@@ -530,8 +530,8 @@ void menu8(phonebook *pb)
     }
 
     printf("Введите id контакта для удаления ссылки: ");
-    unsigned long long id;
-    scanf("%llu", &id);
+    uint64_t id;
+    scanf("%lu", &id);
 
     contact *c = find_by_id(pb, id);
 
@@ -570,8 +570,8 @@ void menu9(phonebook *pb)
     }
 
     printf("Введите id контакта для удаления почты: ");
-    unsigned long long id;
-    scanf("%llu", &id);
+    uint64_t id;
+    scanf("%lu", &id);
 
     contact *c = find_by_id(pb, id);
 
@@ -610,8 +610,8 @@ void menu10(phonebook *pb)
     }
 
     printf("Введите id контакта для удаления телефона: ");
-    unsigned long long id;
-    scanf("%llu", &id);
+    uint64_t id;
+    scanf("%lu", &id);
 
     contact *c = find_by_id(pb, id);
 
@@ -650,8 +650,8 @@ void menu11(phonebook *pb)
     }
 
     printf("Введите id контакта для поиска: ");
-    unsigned long long id;
-    scanf("%llu", &id);
+    uint64_t id;
+    scanf("%lu", &id);
 
     contact *c = find_by_id(pb, id);
 
@@ -678,7 +678,7 @@ void menu12(phonebook *pb)
     char last_name[LAST_NAME_LENGTH];
     scanf("%s", last_name);
 
-    unsigned int count;
+    uint64_t count;
     contact **cs = find_by_last_name(pb, last_name, &count);
 
     if (cs == NULL)
@@ -688,7 +688,7 @@ void menu12(phonebook *pb)
     else
     {
         printf("Найдены контакты:\n");
-        for (unsigned int i = 0; i < count; i++)
+        for (uint64_t i = 0; i < count; i++)
         {
             print_contact(cs[i]);
         }
